Moving stop poller to library function
